src/ml/recommendation-engine.ts
import { z } from 'zod';
import { TensorFlow } from '@tensorflow/tfjs-node';
import { hexPropertyBookingMLRagKnowledgeBase } from './hexproperty-booking-ml-rag-knowledge-base';
import { RecommendationMonitor } from '../monitoring/domain/recommendation-monitor';
import { HealthStatus } from '../monitoring/domain/types';

// STOP: Analyze Current Recommendation Landscape
// Iteration 1: Basic Recommendation Framework

// Domain Models
const UserSchema = z.object({
  id: z.string().uuid(),
  preferences: z.object({
    propertyTypes: z.array(z.enum(['RESIDENTIAL', 'COMMERCIAL', 'VACATION'])),
    priceRange: z.object({
      min: z.number().positive(),
      max: z.number().positive()
    }),
    amenities: z.array(z.string()),
    location: z.object({
      city: z.string(),
      country: z.string()
    })
  }),
  bookingHistory: z.array(z.object({
    propertyId: z.string().uuid(),
    duration: z.number(),
    rating: z.number().min(1).max(5)
  }))
});

const PropertySchema = z.object({
  id: z.string().uuid(),
  type: z.enum(['RESIDENTIAL', 'COMMERCIAL', 'VACATION']),
  price: z.number().positive(),
  amenities: z.array(z.string()),
  location: z.object({
    city: z.string(),
    country: z.string(),
    coordinates: z.object({
      lat: z.number(),
      lon: z.number()
    })
  }),
  bookingHistory: z.array(z.object({
    userId: z.string().uuid(),
    duration: z.number(),
    rating: z.number().min(1).max(5)
  }))
});

// THINK: Design Recommendation Strategy
export class RecommendationEngine {
  private static instance: RecommendationEngine;
  private monitor: RecommendationMonitor;
  private modelId: string;
  private model: any;
  private mlrgKnowledgeBase: typeof hexPropertyBookingMLRagKnowledgeBase;

  private constructor() {
    this.monitor = RecommendationMonitor.getInstance();
    this.modelId = 'recommendation-engine-v1';
    this.initializeModel();
    this.mlrgKnowledgeBase = hexPropertyBookingMLRagKnowledgeBase;

    // Log self-development roadmap on initialization
    console.log('Self-Development Roadmap:', 
      this.mlrgKnowledgeBase.selfDevelopmentRoadmap
    );
  }

  static getInstance(): RecommendationEngine {
    if (!RecommendationEngine.instance) {
      RecommendationEngine.instance = new RecommendationEngine();
    }
    return RecommendationEngine.instance;
  }

  private initializeModel() {
    // Basic collaborative filtering model
    this.model = {
      // Placeholder for ML model initialization
      train: (data: any) => {
        // Simulate model training
        console.log('Training recommendation model');
        return {
          accuracy: Math.random(),
          loss: Math.random()
        };
      },
      predict: (user: z.infer<typeof UserSchema>, properties: z.infer<typeof PropertySchema>[]) => {
        // Basic recommendation logic
        return properties.map(property => ({
          propertyId: property.id,
          score: this.calculateRecommendationScore(user, property)
        })).sort((a, b) => b.score - a.score);
      }
    };
  }

  private calculateRecommendationScore(
    user: z.infer<typeof UserSchema>, 
    property: z.infer<typeof PropertySchema>
  ): number {
    let score = 0;

    // Preference matching
    if (user.preferences.propertyTypes.includes(property.type)) {
      score += 0.3;
    }

    // Price range compatibility
    if (
      property.price >= user.preferences.priceRange.min && 
      property.price <= user.preferences.priceRange.max
    ) {
      score += 0.2;
    }

    // Amenities matching
    const commonAmenities = user.preferences.amenities.filter(amenity => 
      property.amenities.includes(amenity)
    );
    score += commonAmenities.length * 0.1;

    // Location proximity (simplified)
    if (
      user.preferences.location.city === property.location.city ||
      user.preferences.location.country === property.location.country
    ) {
      score += 0.2;
    }

    // Historical booking similarity
    const similarPastBookings = user.bookingHistory.filter(booking => 
      property.bookingHistory.some(propBooking => 
        propBooking.userId === booking.userId
      )
    );
    score += similarPastBookings.length * 0.2;

    return Math.min(score, 1);
  }

  async generateRecommendations(
    user: z.infer<typeof UserSchema>, 
    properties: z.infer<typeof PropertySchema>[]
  ): Promise<any[]> {
    return this.monitor.measurePrediction(this.modelId, async () => {
      try {
        // Record model health before prediction
        await this.monitor.recordModelHealth(this.modelId, HealthStatus.HEALTHY);

        const startTime = Date.now();
        const recommendations = await this.enhanceRecommendationWithMLRG(
          user, 
          properties
        );
        
        // Record accuracy and feature importance
        const accuracy = await this.evaluateAccuracy(recommendations);
        await this.monitor.recordAccuracy(this.modelId, accuracy);
        
        const featureImportance = await this.calculateFeatureImportance();
        await this.monitor.recordFeatureImportance(this.modelId, featureImportance);

        return recommendations;
      } catch (error) {
        // Record unhealthy status on error
        await this.monitor.recordModelHealth(
          this.modelId, 
          HealthStatus.UNHEALTHY,
          error.message
        );
        throw error;
      }
    });
  }

  private async evaluateAccuracy(recommendations: any[]): Promise<number> {
    // Implement accuracy evaluation logic
    return 0.85; // Placeholder
  }

  private async calculateFeatureImportance(): Promise<Record<string, number>> {
    return {
      'location': 0.4,
      'price': 0.3,
      'amenities': 0.2,
      'ratings': 0.1
    };
  }

  async updateModel(newWeights: any): Promise<void> {
    await this.monitor.measurePrediction(this.modelId, async () => {
      try {
        const startTime = Date.now();
        await this.applyModelUpdate(newWeights);
        
        // Record learning rate
        const learningRate = this.calculateLearningRate(newWeights);
        await this.monitor.recordLearningRate(this.modelId, learningRate);
        
        await this.monitor.recordModelHealth(this.modelId, HealthStatus.HEALTHY, 'Model updated successfully');
      } catch (error) {
        await this.monitor.recordModelHealth(
          this.modelId, 
          HealthStatus.UNHEALTHY,
          `Model update failed: ${error.message}`
        );
        throw error;
      }
    });
  }

  private calculateLearningRate(weights: any): number {
    // Implement learning rate calculation
    return 0.01; // Placeholder
  }

  private async enhanceRecommendationWithMLRG(
    user: z.infer<typeof UserSchema>, 
    properties: z.infer<typeof PropertySchema>[]
  ) {
    // Retrieve global and user-specific learnings
    const globalLearnings = await this.mlrgKnowledgeBase.aggregateLearnings(
      'GLOBAL_TREND', 
      {}
    );

    const userGroupLearnings = await this.mlrgKnowledgeBase.aggregateLearnings(
      'USER_GROUP', 
      { 
        ageGroup: this.determineAgeGroup(user),
        incomeLevel: this.determineIncomeLevel(user)
      }
    );

    const locationLearnings = await this.mlrgKnowledgeBase.aggregateLearnings(
      'LOCATION', 
      { 
        city: user.preferences.location.city,
        country: user.preferences.location.country 
      }
    );

    // Integrate MLRG insights into recommendation scoring
    return properties.map(property => {
      const baseScore = this.calculateRecommendationScore(user, property);
      
      const mlrgEnhancements = [
        ...globalLearnings.aggregatedInsights,
        ...userGroupLearnings.aggregatedInsights,
        ...locationLearnings.aggregatedInsights
      ];

      const mlrgScore = mlrgEnhancements.reduce((score, insight) => {
        // Apply MLRG insights to refine recommendation
        switch(insight.key) {
          case 'PROPERTY_TYPE_PREFERENCE':
            return score + (insight.value * insight.confidence);
          case 'PRICE_SENSITIVITY':
            return score + (insight.value * insight.confidence);
          case 'AMENITY_IMPORTANCE':
            return score + (insight.value * insight.confidence);
          default:
            return score;
        }
      }, baseScore);

      return {
        propertyId: property.id,
        baseScore,
        mlrgScore,
        finalScore: (mlrgScore + baseScore) / 2
      };
    }).sort((a, b) => b.finalScore - a.finalScore);
  }

  private determineAgeGroup(user: z.infer<typeof UserSchema>): string {
    // Placeholder method - would typically use more sophisticated logic
    return 'YOUNG_PROFESSIONAL';
  }

  private determineIncomeLevel(user: z.infer<typeof UserSchema>): string {
    // Placeholder method - would typically use more sophisticated logic
    return 'MIDDLE_INCOME';
  }

  async recommendProperties(
    user: z.infer<typeof UserSchema>, 
    properties: z.infer<typeof PropertySchema>[]
  ) {
    const mlrgEnhancedRecommendations = await this.enhanceRecommendationWithMLRG(
      user, 
      properties
    );

    // Record learning from this recommendation session
    await this.recordRecommendationLearning(
      user, 
      mlrgEnhancedRecommendations
    );

    return mlrgEnhancedRecommendations.slice(0, 5).map(rec => ({
      propertyId: rec.propertyId,
      score: rec.finalScore
    }));
  }

  private async recordRecommendationLearning(
    user: z.infer<typeof UserSchema>,
    recommendations: any[]
  ) {
    // Record global learnings
    await this.mlrgKnowledgeBase.addLearning(
      'GLOBAL_TREND',
      {},
      [
        {
          key: 'RECOMMENDATION_DIVERSITY',
          value: recommendations.length,
          confidence: 0.8
        }
      ]
    );

    // Record user group learnings
    await this.mlrgKnowledgeBase.addLearning(
      'USER_GROUP',
      {
        ageGroup: this.determineAgeGroup(user),
        incomeLevel: this.determineIncomeLevel(user)
      },
      recommendations.map((rec, index) => ({
        key: 'RECOMMENDATION_PREFERENCE',
        value: rec.finalScore,
        confidence: 1 - (index * 0.2) // Confidence decreases for lower-ranked recommendations
      }))
    );

    // Record individual user learnings
    await this.mlrgKnowledgeBase.addLearning(
      'INDIVIDUAL_USER',
      { userId: user.id },
      recommendations.map((rec, index) => ({
        key: 'USER_RECOMMENDATION_INTERACTION',
        value: rec.finalScore,
        confidence: 1 - (index * 0.2)
      }))
    );
  }

  private async applyModelUpdate(newWeights: any) {
    // Implement model update logic
  }
}

// Export updated recommendation engine
export const recommendationEngine = RecommendationEngine.getInstance();
